+++ components/Badge.svelte
@@ -1,17 +1,34 @@
 <script>
   export let color = 'gray'
+  export let borderRadius = '50%'
+  export let padding = '.2em'
+  export let bordered = false
+
+  let transparentColor = ''
+
+  $: if (/^#([0-9A-F]{3}){1,2}$/i.test(color)) {
+    color.length == 7 && (transparentColor = color + '12')
+
+    color.length == 4 && (transparentColor = '#' + color.slice(1, 2).repeat(2) + color.slice(2, 3).repeat(2) + color.slice(3, 4).repeat(2) + '12')
+  }
 </script>
 
 <style>
   span {
     min-width: 1.15em;
     min-height: 1.15em;
-    padding: .2em;
+    padding: var(--padding);
+    background-color: var(--theme-color);
+    border-radius: var(--border-radius);
   }
 
-  .background {
-    background-color: var(--theme-color);
+  .bordered {
+    background: linear-gradient(0deg, var(--bg-color), var(--bg-color)), #FFF;
+    border: 1px solid var(--theme-color);
+    color: black;
   }
 </style>
 
-<span class="mdc-typography dib text-align-center white fs-16 br-50 background {$$props.class}" style="--theme-color: {color}" ><slot /></span>
+<span class="mdc-typography dib text-align-center white fs-16 {$$props.class}" class:bordered style="--theme-color: {color}; --bg-color: {transparentColor}; --border-radius: {borderRadius}; --padding: {padding}" >
+  <slot />
+</span>