+++ components/StaticChip.svelte
@@ -1,16 +1,20 @@
+<script>
+  export let bgColor = '#e5e5e5'
+</script>
+
 <style>
-  .chip{
-    background-color: #e5e5e5;
+  .chip {
+    background-color: var(--theme-color);
     height: 36px;
   }
 
-  .chip-content{
+  .chip-content {
     padding-left: 12px;
     padding-right: 12px;
   }
 </style>
 
-<div class="mdc-typography chip black flex justify-center align-items-center mb-1 mr-2 fs-14 br-16px {$$props.class}">
+<div class="mdc-typography chip black flex justify-center align-items-center mb-1 mr-2 fs-14 br-16px {$$props.class}" style="--theme-color: {bgColor}">
   <div class="chip-content flex align-items-center">
     <slot></slot>
   </div>