+++ stories/Badge.stories.svelte
@@ -5,7 +5,9 @@
 
   const args = {
     class: '', //only works for global classes
-    color: 'gray'
+    color: 'gray',
+    borderRadius: '50%',
+    padding: '.2em'
   }
 </script>
 
@@ -26,4 +28,14 @@
 <Story
   name="Blue"
   args={copyAndModifyArgs(args, {color: 'blue'})}
+/>
+
+<Story
+  name="Pill"
+  args={copyAndModifyArgs(args, {borderRadius: '12px', padding: '.1em .5em'})}
+/>
+
+<Story
+  name="Bordered"
+  args={copyAndModifyArgs(args, {borderRadius: '12px', padding: '.1em .5em', color: '#F5731E', bordered: true})}
 />